fix: large component decomposition
